examples/browser-use/requirements.txt
browser-use
langchain-openai
python-dotenv
